+++ src/mpid/common/shm/mpidu_init_shm_alloc.c
@@ -82,7 +82,7 @@ int MPIDU_Init_shm_alloc(size_t len, void **ptr)
 
             mpl_err = MPL_shm_hnd_get_serialized_by_ref(memory->hnd, &serialized_hnd);
             MPIR_ERR_CHKANDJUMP(mpl_err, mpi_errno, MPI_ERR_OTHER, "**alloc_shar_mem");
-            serialized_hnd_size = strlen(serialized_hnd) + 1;   /* add 1 for null char */
+            serialized_hnd_size = (int) strlen(serialized_hnd) + 1;     /* add 1 for null char */
 
             MPIDU_Init_shm_put(serialized_hnd, serialized_hnd_size);
             MPIDU_Init_shm_barrier();
